// lib/api/generated/models/planAssignment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */
import type { AssignmentKind } from './assignmentKind';

export interface PlanAssignment {
  group_ids?: string[];
  kind?: AssignmentKind;
  user_ids?: string[];
}