// lib/api/generated/models/planSkeletonBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */
import type { IntensityModel } from './intensityModel';
import type { PlanAssignment } from './planAssignment';
import type { PlanLevel } from './planLevel';
import type { RpeMode } from './rpeMode';
import type { SkeletonDay } from './skeletonDay';

export interface PlanSkeletonBody {
  assignment?: PlanAssignment;
  days: SkeletonDay[];
  /**
     * @minimum 0
     * @nullable
     */
  deload_every?: number | null;
  include_warmup_block?: boolean;
  include_warmup_sets?: boolean;
  intensity?: IntensityModel;
  is_current?: boolean;
  level?: PlanLevel;
  /** @nullable */
  notes?: string | null;
  rpe_mode?: RpeMode;
  /** @nullable */
  starts_on?: string | null;
  title: string;
  /** @items.minimum 0 */
  weekdays: number[];
  /** @minimum 0 */
  weeks: number;
}