// lib/api/generated/models/skeletonDay.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */
import type { SessionTemplate } from './sessionTemplate';
import type { SkeletonLift } from './skeletonLift';

export interface SkeletonDay {
  accessories?: SkeletonLift[];
  mains?: SkeletonLift[];
  session?: SessionTemplate;
  /** @minimum 0 */
  weekday: number;
}