// lib/api/generated/models/skeletonLift.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */
import type { PctOf } from './pctOf';

export interface SkeletonLift {
  /** @nullable */
  end_pct?: number | null;
  /** @nullable */
  library_id?: string | null;
  /** @nullable */
  load_kg?: number | null;
  name: string;
  pct_of?: null | PctOf;
  /** @nullable */
  start_pct?: number | null;
  work_reps?: string;
  /** @minimum 0 */
  work_sets?: number;
}